refac(zk): move examples to `zk/plonk/examples`
