+++ tachyon/zk/plonk/examples/simple_circuit_unittest.cc
@@ -1,10 +1,10 @@
-#include "tachyon/zk/plonk/circuit/examples/simple_circuit.h"
+#include "tachyon/zk/plonk/examples/simple_circuit.h"
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
-#include "tachyon/zk/plonk/circuit/examples/circuit_test.h"
 #include "tachyon/zk/plonk/circuit/floor_planner/simple_floor_planner.h"
+#include "tachyon/zk/plonk/examples/circuit_test.h"
 #include "tachyon/zk/plonk/halo2/pinned_verifying_key.h"
 #include "tachyon/zk/plonk/keys/proving_key.h"
