+++ tachyon/zk/plonk/vanishing/BUILD.bazel
@@ -173,10 +173,10 @@ tachyon_cc_unittest(
         "//tachyon/zk/base/entities:verifier_base",
         "//tachyon/zk/expressions:expression_factory",
         "//tachyon/zk/expressions/evaluator/test:evaluator_test",
-        "//tachyon/zk/plonk/circuit/examples:circuit_test",
-        "//tachyon/zk/plonk/circuit/examples:simple_circuit",
         "//tachyon/zk/plonk/circuit/floor_planner:simple_floor_planner",
         "//tachyon/zk/plonk/constraint_system",
+        "//tachyon/zk/plonk/examples:circuit_test",
+        "//tachyon/zk/plonk/examples:simple_circuit",
         "//tachyon/zk/plonk/halo2:pinned_verifying_key",
         "//tachyon/zk/plonk/halo2:prover_test",
         "//tachyon/zk/plonk/keys:proving_key",
